+++ Sources/IRGen/EVM.swift
@@ -5,10 +5,12 @@
 //  Created by Franklin Schrans on 5/3/18.
 //
 
+import BigInt
+
 /// Constants related to EVM.
 enum EVM {
   /// The size of an EVM storage location, in bytes.
-  static var wordSize: Int {
+  static var wordSize: BigInt {
     return 32
   }
 }
